Csv importing (#8)
* Organises Models and namespaces

* Adds basic UserData model

* Set out function with interfaces and stub implementantions

* Slight tweak to interface naming and description to make it much more generic and less tightly coupled to the current solution

* Adds docblock

* Adds DI for the services into the function, and gets the HistoryService compiling

* Fixes parameterised test type

* Adds documentation to test helper methods

* Fixes incorrect parameters in docblocks for helper overloads

* Alters public constant casing from SCREAMING_SNAKE_CASE (love that name) to PascalCase as per microsoft naming convention. (...and to remove IDE warnings)

* Converts select callback to method group syntax for readablity

* Initialises DB Tables

Initial implementation of UserRepository with passing test.

Extracts SQL commands from strings to file imports to stop IDE complaining about SQL dialect

* Updates the user repository interface with required methods.

Adds Async suffix to asynchronous methods

* Updates repository to meet interface

* Failing test to test user creation with/without entraid

* Renames cloumn names, passes tests

* Adds failing test to test unique constraint on entraid

* Updates repository method to catch and return expected exception

* fixes build warning of unused variable

* Runs dotnet format on whole repo

* Adds documentation to user repository.

* Refactors to extract sql queries into constants class. adds documentation for each query

* Adds documentation to DI constructor, and class as a whole.

* Creates tests to check for invalid entraIds during user creation

* Passes tests by restricting user creation to only users providing a valid entraId

* Adds a failing test exposing a bug where SQL call is returning affected row count rather than id value

* Updates repository method to fix bug, adds documentation and link

* Refactors tests to extract guid creation logic, fixes grammar in comment

* Creates and documents GetRunEventsByUserIdAsync method to query and return previously saved run events.

* Refactors Repostitory tests to use regions

* Refactors with passing tests to extract creation of fake RunEvents to helper methods

* Parameterises rowcount test

* Adds multiple test cases to affected rows test

* Adds failing test for invalid userid

* Adds test to check runevent is in the past

* Adds validation method to validate RunEvent before adding to DB

* Updates failing test for future date to be less specific as a sql error

* Extracts fake methods into shared Fakes class

* Adds validator, validator interface and failing failure path tests

* Updates repository test to test null event entry

* Updates failing pipeline by using direct repo reference and branch pinning

* Changes to tests.  Adds test for null events collection, and adds tests for graceful discarding of null events rather than a throw

* Failing test for getrunevents indicates parse error for duration storage in sqllite

* Fixes internal DB model, adjust dapper's mapping on retrieval.

* Fixes RunEvent validator tests, passes failing tests and fixes error in interface

* Adds and Fixes documentation in Validator

* uses collection initialiser

* Parameterises test to test multiple event retrieval.

* Adds test to test multiple events added to multiple users

* Extracts CreateRunEvents method to return a collection of default run events

* Minor test refactorings

* Adds a multiple insertion test to test a single user data retrieval request from multiple insertions

* Adds test for invalid userid to return an empty collection rather than throwing

* Refactors database preseeding logic into a test helper

* Runs dotnet format on whole repo

* Adds failing tests for GetUserDataByUserId repository method

* WIP - adding repository user lookup sql and repository methods to pass failing tests

+++ tests/RunSuggestion.Core.Tests/Transformers/CsvToRunHistoryTransformerTests.cs
@@ -1,6 +1,6 @@
 using RunSuggestion.Core.Interfaces;
-using RunSuggestion.Core.Models;
-using RunSuggestion.Core.Models.DataSources.TrainingPeaks;
+using RunSuggestion.Core.Models.Runs;
+using RunSuggestion.Core.Models.Runs.DataSources.TrainingPeaks;
 using RunSuggestion.Core.Services;
 using RunSuggestion.Core.Tests.TestHelpers;
 using RunSuggestion.Core.Transformers;
@@ -81,7 +81,7 @@ public class CsvToRunHistoryTransformerTests
     [InlineData(1D)]
     [InlineData(10D)]
     [InlineData(100D)]
-    public void ConvertToRunHistory_WhenPassedValidDuration_ParsesDurationCorrectly(int duration)
+    public void ConvertToRunHistory_WhenPassedValidDuration_ParsesDurationCorrectly(double duration)
     {
         // Arrange
         TimeSpan expectedDuration = TimeSpan.FromHours(duration);
